Update README.md
Even that I am retired, (born is 1940) still interested in AI Tools and its application in businees and health. 
I start to use chatGPT in 2023 febr. 
Now stiil using chatGPT, a bit Claude and Perplexity.
My actual PERSONAL PROJECT IS TO DEVELOP a stable envirenment with Github assistance to use such AI Tools as MONAILABEL for studying my own CT Studies. 
The other sphere is to use / develop envirenment for  using 
AI tools to assist in the preparation of our condominium's annual budget and renovation plans.